Updated project format

+++ src/main/scala/fpinscala/samples/chapter02/StackOverflow.scala
@@ -1,3 +1,18 @@
+// Copyright (C) 2016 the original author or authors.
+// See the LICENCE.txt file distributed with this work for additional
+// information regarding copyright ownership.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 package fpinscala.samples.chapter02
 
 // WARNING: Long return types will likely cause overflow computing sums
@@ -8,7 +23,7 @@ package fpinscala.samples.chapter02
 object StackOverflow {
 
   // sum from math (Gauss formula)
-  def sum(n: Int): Long = n * (n + 1) / 2
+  def sum(n: Int): Long = (n * (n + 1) / 2).toLong
 
   // recursive sum of integer from 0 to n
   def sum1(n: Int): Long = n match {
